+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 ### Added
+- Compliant fields are recorded and are now part of the Object Standards grade output
 
 ### Changed